refactor (NuGettier.Upm): move implementation for StringStringDictionary into its own file
reason: re-use by NPM protocol
