+++ NuGettier.Upm/PackageJson/PackageJson.cs
@@ -6,20 +6,6 @@ namespace NuGettier.Upm;
 
 public class PackageJson
 {
-    public sealed class StringStringDictionary : Dictionary<string, string>
-    {
-        public StringStringDictionary()
-            : base() { }
-
-        public StringStringDictionary(StringStringDictionary reference)
-            : this(reference as Dictionary<string, string>) { }
-
-        public StringStringDictionary(Dictionary<string, string> reference)
-            : base(reference) { }
-
-        public StringStringDictionary(IEnumerable<KeyValuePair<string, string>> reference)
-            : base(reference) { }
-    }
 
     public string Name = String.Empty;
     public string name
